okvs-range: better specification of the options
The options of okvs-range must be applied in a certain order to make
sens (and be portable). They must be applied in the order they appear
in the specification:

- reverse?
- offset
- limit

If reverse? is #f, skip it. If offset is #f skip it. If limit is #f
return all-the-things (in a generator).

There is an optimisation looming depending depending on the backend.
If reverse? is #t then the key space can be navigated from back to
front. if offset is a number, some keys can be skipped. Limit requires
a counter inside the generator. The idea is to keep space usage low
(and spread it over time via the generator (unlike the sample
implementation (as of yet))).